BUG: sparse.linalg: fix overflow in `spsolve` due to 32-bit int usage
Fixed the signature of doubleMalloc and particularly doubleCalloc:
possibly malloc needs to allocate >= 2^31 bytes.

Fixed the signature of doubleMalloc and doubleCalloc in the
implementation file: `SUPERLU_MALLOC` now passes the correct value to
`void *superlu_python_module_malloc(size_t size)` which internally calls
`malloc(size)`.

`work = doubleCalloc(n*nrhs)` now executes for right-hand sides B of
size 46679 by 46680 filled with doubles, and some `(size_t)` casts that
fix proper indexing into buffers such as
`work_col = &work[(size_t)j * (size_t)n]`.

+++ scipy/sparse/linalg/_dsolve/SuperLU/SRC/dmemory.c
@@ -672,22 +672,22 @@ dallocateA(int n, int nnz, double **a, int **asub, int **xa)
 }
 
 
-double *doubleMalloc(int n)
+double *doubleMalloc(size_t n)
 {
     double *buf;
-    buf = (double *) SUPERLU_MALLOC((size_t)n * sizeof(double)); 
+    buf = (double *) SUPERLU_MALLOC(n * (size_t) sizeof(double));
     if ( !buf ) {
 	ABORT("SUPERLU_MALLOC failed for buf in doubleMalloc()\n");
     }
     return (buf);
 }
 
-double *doubleCalloc(int n)
+double *doubleCalloc(size_t n)
 {
     double *buf;
-    register int i;
+    register size_t i;
     double zero = 0.0;
-    buf = (double *) SUPERLU_MALLOC((size_t)n * sizeof(double));
+    buf = (double *) SUPERLU_MALLOC(n * (size_t) sizeof(double));
     if ( !buf ) {
 	ABORT("SUPERLU_MALLOC failed for buf in doubleCalloc()\n");
     }
